requirements.txt
alpaca-trade-api==3.1.1
polygon-api-client==1.13.0
openai==1.12.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.2.0
numpy==1.26.3
scipy==1.12.0
ta==0.11.0
APScheduler==3.10.4

